# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.CreatePackageRequest:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      question_ids:
        items:
          type: string
        type: array
      status:
        type: string
    required:
    - code
    - name
    - question_ids
    type: object
...
  entity.DeleteUserRequest:
    properties:
      reason:
        type: string
    required:
    - reason
    type: object
...
  entity.GetPackageListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.PackageListItem'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
...
  entity.PackageDetailResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      questions:
        items:
          $ref: '#/definitions/entity.PackageQuestionInfo'
        type: array
      status:
        type: string
      updated_at:
        type: string
      verification_status:
        type: string
    type: object
  entity.PackageListItem:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      question_count:
        type: integer
      status:
        type: string
      verification_status:
        type: string
    type: object
  entity.PackageQuestionInfo:
    properties:
      question_code:
        type: string
      question_id:
        type: string
      sort_order:
        type: integer
    type: object
...
  entity.RoleInfo:
    properties:
      code:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  entity.UpdateUserRequest:
    properties:
      email:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      nip:
        type: string
      password:
        description: kosong = tidak diubah
        type: string
      role_ids:
        items:
          type: string
        type: array
      satker_id:
        type: string
      username:
        type: string
    type: object
  entity.UserDetailResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      nip:
        type: string
      roles:
        items:
          $ref: '#/definitions/entity.RoleInfo'
        type: array
      satker_id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  helper.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend service (hexagon/clean architecture).
  title: Widyaprada Backend API
  version: "1.0"
paths:
...
  /api/v1/question-packages:
    get:
      consumes:
      - application/json
      description: Get paginated list of question packages with filters
      parameters:
      - description: Search
        in: query
        name: q
        type: string
      - description: Draft, Aktif, all
        in: query
        name: status
        type: string
      - description: Belum, Sudah, all
        in: query
        name: status_verifikasi
        type: string
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.GetPackageListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List question packages
      tags:
      - question-packages
    post:
      consumes:
      - application/json
      parameters:
      - description: Create package request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.CreatePackageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.PackageDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create question package
      tags:
      - question-packages
...
  /api/v1/users/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete user by ID (reason required)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Delete reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.DeleteUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserDetailResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user detail
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update user request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.UserDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
swagger: "2.0"